human/requirements.txt
amqp==2.2.2
awsebcli==3.12.4
billiard==3.5.0.3
botocore==1.9.16
celery==4.1.0
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
djangorestframework==3.7.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
pathspec==0.5.5
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
websocket-client==0.47.0
wincertstore==0.2
